requirements.txt
apache-airflow-providers-sqlite
pandas